Reformatted to doc, helpers and metrics.go
