+++ beacon-chain/blockchain/forkchoice/doc.go
@@ -1,0 +1,4 @@
+/*
+	Package forkchoice Add details on how forkchoice works.
+*/
+package forkchoice
