+++ beacon-chain/blockchain/forkchoice/metrics.go
@@ -1,0 +1,1 @@
+package forkchoice
